# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
        "/guests": {
            "get": {
                "tags": [
                    "Гости"
                ],
                "summary": "Получить список гостей с пагинацией",
                "operationId": "04216ccbb20f5c202a482672144feb01",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице (по умолчанию 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы (по умолчанию 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список гостей с метаданными пагинации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GuestResource"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            },
            "post": {
                "tags": [
                    "Гости"
                ],
                "summary": "Создать нового гостя",
                "operationId": "a61d35217f147b5a0f768836339f01cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Иван"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Иванов"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "primer@mail.ru"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+78002000600"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Гость успешно создан"
                    },
                    "500": {
                        "description": "Ошибка при создании гостя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Не удалось создать гостя"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/guests/{id}": {
            "get": {
                "tags": [
                    "Гости"
                ],
                "summary": "Получить информацию о госте по id",
                "operationId": "c9a5aa22f95e71a6ff0c593fe48d7034",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id гостя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о госте",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuestResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Гость не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Гость не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Произошла ошибка на сервере"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Гости"
                ],
                "summary": "Обновить информацию о госте",
                "operationId": "3828d5f48402c49c6885b23dda995f3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id гостя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Иван"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Иванов"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "primer@mail.ru"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+78002000600"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "RU"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Гость успешно обновлен"
                    },
                    "404": {
                        "description": "Гость не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Гость не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при обновлении гостя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Не удалось обновить гостя"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Гости"
                ],
                "summary": "Удалить гостя",
                "operationId": "e2c1ad7578158f7097f770d60908e399",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID гостя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Гость успешно удален"
                    },
                    "404": {
                        "description": "Гость не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Гость не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при удалении гостя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Не удалось удалить гостя"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "GuestResource": {
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "Elmore"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Ruecker"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "macie.hickle@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+12287197893"
                    },
                    "country": {
                        "type": "string",
                        "example": "TD"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Гости",
            "description": "Гости"
        }
    ]
}